Change int to long

+++ app/src/main/java/com/example/application/detail/viewmodels/WeatherDetailViewModel.kt
@@ -48,7 +48,7 @@ class WeatherDetailViewModel(private val weatherService: WeatherService) : ViewM
                 Timber.i("response successful weather items count for ${weatherResponse.city.name}: ${weatherResponse.list.size}")
 
                 weatherResponse.list.forEach { weatherItem ->
-                    val unixTimestamp = weatherItem.dt.toLong().times(1000)
+                    val unixTimestamp = weatherItem.dt.times(1000)
                     val currentDay = dateFormatDay.format(unixTimestamp)
 
                     if (selectedDay == currentDay) {
